+++ crates/forge/tests/cli/test_cmd/mod.rs
@@ -22,18 +22,7 @@ mod symbolic_conformance;
 mod symbolic_engine_capabilities;
 mod symbolic_helpers;
 mod symbolic_limits;
-mod symbolic_parity_crytic;
-mod symbolic_parity_devdacian;
-mod symbolic_parity_echidna;
-mod symbolic_parity_halmos;
-mod symbolic_parity_hevm;
-mod symbolic_parity_ityfuzz;
-mod symbolic_parity_kevm;
-mod symbolic_parity_manticore;
-mod symbolic_parity_medusa;
-mod symbolic_parity_openzeppelin;
-mod symbolic_parity_scribble;
-mod symbolic_parity_swc;
+mod symbolic_parity;
 mod table;
 mod trace;
